// src/types/tooltipPositions.ts
import TooltipPosition from "./tooltipPosition"

type TooltipPositions = {
    'left': [ TooltipPosition, TooltipPosition, TooltipPosition, TooltipPosition]
    'top': [ TooltipPosition, TooltipPosition, TooltipPosition, TooltipPosition]
    'right': [ TooltipPosition, TooltipPosition, TooltipPosition, TooltipPosition]
    'bottom': [ TooltipPosition, TooltipPosition, TooltipPosition, TooltipPosition]
}

export default TooltipPositions